requirements.txt
aws-cdk-lib>=2.78.0
constructs>=10.0.0,<11.0.0
boto3
cdk-aws-lambda-powertools-layer
Pillow==9.4.0